+++ src/xulbux/_consts_.py
@@ -85,7 +85,7 @@ class ANSI:
     ]
 
     codes_map: dict[str | tuple[str], int] = {
-        ###################### RESETS ######################
+        ################# SPECIFIC RESETS ##################
         "_": 0,
         ("_bold", "_b"): 22,
         ("_dim", "_d"): 22,
@@ -97,7 +97,7 @@ class ANSI:
         ("_strikethrough", "_s"): 29,
         ("_color", "_c"): 39,
         ("_background", "_bg"): 49,
-        ################### TEXT FORMATS ###################
+        ################### TEXT STYLES ####################
         ("bold", "b"): 1,
         ("dim", "d"): 2,
         ("italic", "i"): 3,
@@ -106,7 +106,7 @@ class ANSI:
         ("hidden", "hide", "h"): 8,
         ("strikethrough", "s"): 9,
         ("double-underline", "du"): 21,
-        ############## DEFAULT CONSOLE COLORS ##############
+        ################## DEFAULT COLORS ##################
         "black": 30,
         "red": 31,
         "green": 32,
@@ -115,7 +115,7 @@ class ANSI:
         "magenta": 35,
         "cyan": 36,
         "white": 37,
-        ########## BRIGHT DEFAULT CONSOLE COLORS ###########
+        ############## BRIGHT DEFAULT COLORS ###############
         "br:black": 90,
         "br:red": 91,
         "br:green": 92,
@@ -124,7 +124,7 @@ class ANSI:
         "br:magenta": 95,
         "br:cyan": 96,
         "br:white": 97,
-        ######## DEFAULT CONSOLE BACKGROUND COLORS #########
+        ############ DEFAULT BACKGROUND COLORS #############
         "bg:black": 40,
         "bg:red": 41,
         "bg:green": 42,
@@ -133,7 +133,7 @@ class ANSI:
         "bg:magenta": 45,
         "bg:cyan": 46,
         "bg:white": 47,
-        ##### BRIGHT DEFAULT CONSOLE BACKGROUND COLORS #####
+        ######### BRIGHT DEFAULT BACKGROUND COLORS #########
         "bg:br:black": 100,
         "bg:br:red": 101,
         "bg:br:green": 102,